webapi/webapi-messaging-w3c-tests/messaging/SmsMessage.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Shentu.Jiazhen <jiazhenx.shentu@intel.com>
        Jianfeng.Xu <jianfengx.xu@intel.com>

-->

<meta charset='utf-8'>
<title>Messaging Test: SmsMessage</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://www.w3.org/2012/sysapps/messaging/#smsmessage-interface">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="support/support.js"></script>
<div id="log"></div>
<script>

var testNumber = "5556";
var filter = {
  type: "sms",
};
var filterOption = {
  sortBy: "date",
  sortBy: "descending",
  limit: 3
};
var t1 = async_test("Check if the messageID of MmsMessage is the identifier of the message", {timeout: 2000});

var t2 = async_test("Check if the type of SmsMessage is the type of message", {timeout: 2000});

var t3 = async_test("Check if the serviceID of SmsMessage is the messaging service id used to send / receive this message", {timeout: 2000});

var t4 = async_test("Check if the from of SmsMessage is the sender of the message", {timeout: 2000});

var t5 = async_test("Check if the timestamp of SmsMessage is the time the message reached the Short Message Center", {timeout: 2000});

var t6 = async_test("Check if the read of SmsMessage is 'true' if the message has been marked as read", {timeout: 2000});

var t7 = async_test("Check if the read of SmsMessage is 'false' if the message has not been marked as read", {timeout: 2000});

var t8 = async_test("Check if the to of SmsMessage is the recipient of the message", {timeout: 2000});

var t9 = async_test("Check if the body of SmsMessage is the text of the SMS message", {timeout: 2000});

var t10 = async_test("Check if the state of SmsMessage is in the states of SmsState", {timeout: 2000});

var t11 = async_test("Check if the deliveryStatus of SmsMessage is in the states of DeliveryStatus", {timeout: 2000});

var t12 = async_test("Check if the deliveryTimestamp of SmsMessage for sent messages the delivery date as reported in the TP-Discharge-Time (TP DT) parameter included in the SMS-STATUS-REPORT Protocol Data Unit", {timeout: 2000});

var t13 = async_test("Check if the deliveryTimestamp of SmsMessage is null if there is no positive knowledge about the delivery of the message", {timeout: 2000});

var t14 = async_test("Check if the deliveryTimestamp of SmsMessage is null for received messages", {timeout: 2000});

var t15 = async_test("Check if the messageClass of SmsMessage is in the states of MessageClass", {timeout: 2000});

setup(
  function () {
    message.sms.send(testNumber, "Hello Xwalk!");
  }
)

t1.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t1.step(function () {
        assert_true(msg.element.length > 0, "No message was finded");
        assert_true(!!msg.element[0].messageID, "The messageID of SmsMessage");
      });
      t1.done();
    },
    function (error) {
      t1.step(function () {
        assert_unreached(error.message);
      });
      t1.done();
    }
  );
});

t2.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t2.step(function () {
        assert_true(msg.element.length > 0, "No message was finded");
        assert_equals(msg.element[0].type, "sms", "The type of SmsMessage");
      });
      t2.done();
    },
    function (error) {
      t2.step(function () {
        assert_unreached(error.message);
      });
      t2.done();
    }
  );
});

t3.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t3.step(function () {
        assert_true(msg.element.length > 0, "No message was finded");
        assert_equals(msg.element[0].serviceID, "sim0", "The serviceID of SmsMessage");
      });
      t3.done();
    },
    function (error) {
      t3.step(function () {
        assert_unreached(error.message);
      });
      t3.done();
    }
  );
});

t4.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t4.step(function () {
        assert_true(msg.element.length > 0, "No message was finded");
        assert_true(!!msg.element[0].from, "The from of SmsMessage");
      });
      t4.done();
    },
    function (error) {
      t4.step(function () {
        assert_unreached(error.message);
      });
      t4.done();
    }
  );
});

t5.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t5.step(function () {
        assert_true(msg.element.length > 0, "No message was finded");
        assert_true(!!msg.element[0].timestamp, "The timestamp of SmsMessage");
      });
      t5.done();
    },
    function (error) {
      t5.step(function () {
        assert_unreached(error.message);
      });
      t5.done();
    }
  );
});

t6.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t6.step(function () {
        assert_true(msg.element.length > 0, "No message was finded");
        assert_true(!!msg.element[0].read, "SmsMessage.read");
      });
      t6.done();
    },
    function (error) {
      t6.step(function () {
        assert_unreached(error.message);
      });
      t6.done();
    }
  );
});

t7.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t7.step(function () {
        assert_true(msg.element.length > 0, "No message was finded");
        assert_equals(typeof(msg.element[0].read), "boolean", "SmsMessage.read");
      });
      t7.done();
    },
    function (error) {
      t7.step(function () {
        assert_unreached(error.message);
      });
      t7.done();
    }
  );
});

t8.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t8.step(function () {
        assert_true(msg.element.length > 0, "No message was finded");
        assert_equals(msg.element[0].to, "5556", "The to of SmsMessage");
      });
      t8.done();
    },
    function (error) {
      t8.step(function () {
        assert_unreached(error.message);
      });
      t8.done();
    }
  );
});

t9.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t9.step(function () {
        assert_true(msg.element.length > 0, "No message was finded");
        assert_equals(msg.element[0].body, "Hello Xwalk!", "The body of SmsMessage");
      });
      t9.done();
    },
    function (error) {
      t9.step(function () {
        assert_unreached(error.message);
      });
      t9.done();
    }
  );
});

t10.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t10.step(function () {
        assert_true(msg.element.length > 0, "No message was finded");
        assert_in_array(msg.element[0].state, ["received", "draft", "sending", "sent", "failed"], "The state of SmsMessage");
      });
      t10.done();
    },
    function (error) {
      t10.step(function () {
        assert_unreached(error.message);
      });
      t10.done();
    }
  );
});

t11.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t11.step(function () {
        assert_true(msg.element.length > 0, "No message was finded");
        assert_in_array(msg.element[0].deliveryStatus, ["success", "pending", "error", "not-applicable"], "The deliveryStatus of SmsMessage");
      });
      t11.done();
    },
    function (error) {
      t11.step(function () {
        assert_unreached(error.message);
      });
      t11.done();
    }
  );
});

t12.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t12.step(function () {
        assert_true(msg.element.length > 0, "No message was finded");
        assert_true(!!msg.element[0].deliveryTimestamp, "The deliveryTimestamp of SmsMessage");
      });
      t12.done();
    },
    function (error) {
      t12.step(function () {
        assert_unreached(error.message);
      });
      t12.done();
    }
  );
});

t13.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t13.step(function () {
        assert_true(msg.element.length > 0, "No message was finded");
        assert_true(!!msg.element[0].deliveryTimestamp, "The deliveryTimestamp of SmsMessage");
      });
      t13.done();
    },
    function (error) {
      t13.step(function () {
        assert_unreached(error.message);
      });
      t13.done();
    }
  );
});

t14.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t14.step(function () {
        assert_true(msg.element.length > 0, "No message was finded");
        assert_equals(msg.element[0].deliveryTimestamp, null, "The deliveryTimestamp of SmsMessage");
      });
      t14.done();
    },
    function (error) {
      t14.step(function () {
        assert_unreached(error.message);
      });
      t14.done();
    }
  );
});

t15.step(function () {
  assert_true(!!message, "The define of message");
  message.findMessages(filter, filterOption).then(
    function (msg) {
      t15.step(function () {
        assert_true(msg.element.length > 0, "No message was finded");
        assert_in_array(msg.element[0].messageClass, ["class-0", "class-1", "class-2", "class-3", "normal"], "The messageClass of SmsMessage");
      });
      t15.done();
    },
    function (error) {
      t15.step(function () {
        assert_unreached(error.message);
      });
      t15.done();
    }
  );
});

</script>
